% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_alcohol.R
\name{clean_alcohol}
\alias{clean_alcohol}
\title{Clean Alcohol Variables}
\usage{
clean_alcohol(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all alcohol related variables.
}
