% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_benefits.R
\name{clean_benefits}
\alias{clean_benefits}
\title{Clean Benefit Variables}
\usage{
clean_benefits(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all variables related to social security benefit receipt and monthly
income received from social security benefits.
}
