% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demographic.R
\name{clean_demographic}
\alias{clean_demographic}
\title{Clean Demographic Variables}
\usage{
clean_demographic(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Reads and does basic cleaning on the UKHLS demographic variables -
age, sex, ethnicity, region and urban vs rural.
}
