% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_econ_status.R
\name{clean_econ_status}
\alias{clean_econ_status}
\title{Clean Labour Market}
\usage{
clean_econ_status(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Generate three variables indicating economic status at different levels of detail,
and earnings variables. Variables with "pay" in the name relate to earnings from employment,
"earnings" refers to employed and self employed earnings combined.
}
