% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hours_earn.R
\name{clean_hours_earn}
\alias{clean_hours_earn}
\title{Clean Hours and Earnings Variables}
\usage{
clean_hours_earn(data = NULL)
}
\description{
Reads and does basic cleaning on the UKHLS demographic variables -
age, gender, ethnicity, region and urban vs rural.
}
