% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_income.R
\name{clean_income}
\alias{clean_income}
\title{Clean Income Variables}
\usage{
clean_income(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Reads and does basic cleaning on the UKHLS income variables. Total income (gross and net),
division of net income into labour income, miscellaneous income, private benefit income,
investment income, pension income, social benefit income.
}
