% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_labour_market.R
\name{clean_labour_market}
\alias{clean_labour_market}
\title{Clean Labour Market}
\usage{
clean_labour_market(data = NULL, inflation = ukhlsclean::cpih)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}

\item{inflation}{Data table. CPIH inflation data input for real-terms adjustments.}
}
\description{
Generate three variables indicating economic status at different levels of detail,
and earnings variables. Variables with "pay" in the name relate to earnings from employment,
"earnings" refers to employed and self employed earnings combined.
}
