% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_smoke.R
\name{clean_smoke}
\alias{clean_smoke}
\title{Clean Smoking Variables}
\usage{
clean_smoke(data = NULL)
}
\description{
Clean all variables related to smoking behaviour - current smoking status, smoking history, and
level of cigarette consumption for current smokers.
}
