% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_survey_weights.R
\name{clean_survey_weights}
\alias{clean_survey_weights}
\title{Clean Survey Weights}
\usage{
clean_survey_weights(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Produce consistent variables containing cross-sectional weights for the purpose of
producing representative descriptive statistics by wave.
}
