% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_work.R
\name{clean_work}
\alias{clean_work}
\title{Clean Work}
\usage{
clean_work(data = NULL, calendar_year = FALSE)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}

\item{calendar_year}{Logical - TRUE when the code is processing calendar year data (defaults to FALSE).}
}
\description{
Generate variables (other than pay) relating to work, including hours of work for both
employed and self-employed, and industry / occupation.
}
