% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_youth.R
\name{clean_youth}
\alias{clean_youth}
\title{Clean Youth data Variables}
\usage{
clean_youth(youth_data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Produce clean versions of variables measuring youth smoking outcomes.
}
