% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoefficientMatrix}
\alias{CoefficientMatrix}
\title{EQ-5D Coefficient Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 36 rows and 10 columns.
}
\usage{
CoefficientMatrix
}
\description{
Matrix of coefficients used to calculate EQ-5D domains from the SF-12.
\href{https://www.herc.ox.ac.uk/downloads/downloads-supporting-material-1/sf-12-responses-and-eq-5d-utility-values}{SF-12 Responses and EQ-5D Utility Values}
}
\keyword{datasets}
