% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_waves.R
\name{combine_waves}
\alias{combine_waves}
\title{Combine waves of UKHLS}
\usage{
combine_waves(data_list, ukhls = TRUE)
}
\arguments{
\item{data_list}{List of data tables to combine.}

\item{ukhls}{TRUE if `data_list` contains UKHLS waves}
}
\description{
Combines years of data when provided as a list of data tables.
}
