% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Monthly Consumer Price Index}
\format{A data frame with 351 observations and 3 variables.
\describe{
    \item{year}{}
    \item{month}{}
    \item{cpi_value}{}
}}
\usage{
cpi
}
\description{
Returns the CPI inflation index on a monthly basis (currently set to
re-base the index to March 2020, the most recent month) - data are obtained from the
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/l522/mm23}{ONS website}
last accessed 19/06/2020.
}
\keyword{datasets}
