% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpih}
\alias{cpih}
\title{Monthly Consumer Price Index}
\format{
A data frame with 442 observations and 3 variables.
\describe{
    \item{year}{}
    \item{month}{}
    \item{index}{}
}
}
\usage{
cpih
}
\description{
Monthly CPIH inflation series with a base month of January 2026. Data are obtained from the
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/l522/mm23}{ONS website}
and the data were last updated 20-Feb-2026. Data are available up to Jan-2026.
}
\keyword{datasets}
