% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d}
\alias{eq5d}
\title{EQ-5D Algorithm}
\usage{
eq5d(data, matrix = ukhlsclean::CoefficientMatrix, seed = 0)
}
\arguments{
\item{data}{Data table. UKHLS data containing SF-12 variables.}

\item{matrix}{Matrix. Matrix of coefficients to use in the algorithm}

\item{seed}{Integer. Random number seed}
}
\value{

}
\description{
Apply the algorithm developed by Gray A, Rivero-Arias O, Clarke P (2006) to estimate
EQ-5D utility values from SF-12 responses. Algorithm materials obtained from
\href{https://www.herc.ox.ac.uk/downloads/downloads-supporting-material-1/sf-12-responses-and-eq-5d-utility-values}{SF-12 Responses and EQ-5D Utility Values}
}
\examples{

\dontrun{

}

}
