% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_scaling_factor.R
\name{gen_scaling_factor}
\alias{gen_scaling_factor}
\title{Generate Population Scaling Factor}
\usage{
gen_scaling_factor(data, pop_data = ukhlsclean::PopulationCounts)
}
\arguments{
\item{data}{data table - merged data generated from different cleaning modules in `ukhls_clean_global`}

\item{pop_data}{data table - population counts by single year of age, sex, and UK nation (defaults to `PopulationCounts`)}
}
\value{
data table
}
\description{
Merge ONS mid-year population estimates into the calendar year data to generate a population scaling factor
for calculating estimates of total population statistics. The scaling factor is calculated by summing the total
weighted number of individuals in each sex, country, and single-year age cell in the sample and dividing the total
population by this figure, to get a number of total individuals in the population represented by each individual in the
sample.
}
\examples{

\dontrun{

}
}
