% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_data_clean.R
\name{global_data_clean}
\alias{global_data_clean}
\title{Clean UKHLS data}
\usage{
global_data_clean(
  data,
  ages = 16:89,
  country = NULL,
  complete_vars = NULL,
  calendar_year = FALSE,
  inflation = ukhlsclean::cpih
)
}
\arguments{
\item{data}{Data table - the combined Understanding Society dataset for one wave.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to NULL - all ages).}

\item{country}{Character - country to produce data for. One of c("england","wales","scotland","northern_ireland"). Defaults to NULL which includes all UK.}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to NULL - keep all observations).}

\item{calendar_year}{Logical - TRUE when the code is processing calendar year data (defaults to FALSE).}

\item{inflation}{Data table. Inflation data input for real-terms adjustments. Defaults to CPIH.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults NULL - keep all variables).}
}
\value{
Returns a new set of variables
}
\description{
A wrapper function for applying all of the cleaning functions, selecting the
desired variables/observations for the analysis, and specifying complete case
restrictions.
}
