% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_income_data.R
\name{process_income_data}
\alias{process_income_data}
\title{Clean Unearned-Income File}
\usage{
process_income_data(data.income = NULL)
}
\arguments{
\item{data.income}{Data table. Understanding Society income file.}
}
\description{
Perform basic cleaning on the income data file to process benefit income
data and reshape into a single row per individual.
}
