% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpi}
\alias{rpi}
\title{Monthly Retail Price Index}
\format{
A data frame with 929 observations and 3 variables.
\describe{
    \item{year}{}
    \item{month}{}
    \item{index}{}
}
}
\usage{
rpi
}
\description{
Monthly RPI inflation series with a base month of January 2026. Data are obtained from the
\href{https://www.ons.gov.uk/generator?format=xls&uri=/economy/inflationandpriceindices/timeseries/cdko/mm23}{ONS website}
and the data were last updated 20-Feb-2026. Data are available up to Jan-2026.
}
\keyword{datasets}
