% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select variables and apply filters}
\usage{
select_data(
  data,
  ages = 16:89,
  waves = 1:11,
  keep_vars = NULL,
  complete_vars = c("age")
)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{waves}{Integer vector - the waves of the UKHLS to retain (defaults to all - 1 to 11).}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to year and age).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year and age).}
}
\value{
Returns a reduced version of data
}
\description{
Selects the variables required for analysis and selects only the rows without missing data
for specified variables.
}
\examples{

\dontrun{

data <- select_data(data, keep_vars = c("age", "gender", "region"))

}

}
