% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_alcohol.R
\name{ukhls_clean_alcohol}
\alias{ukhls_clean_alcohol}
\title{Clean Alcohol Variables}
\usage{
ukhls_clean_alcohol(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all alcohol related variables.
}
