% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_benefit.R
\name{ukhls_clean_benefit}
\alias{ukhls_clean_benefit}
\title{Clean Benefit Variables}
\usage{
ukhls_clean_benefit(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all variables related to benefits -
}
