% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_health_condition.R
\name{ukhls_clean_health_conditions}
\alias{ukhls_clean_health_conditions}
\title{Clean Health and Caring Variables}
\usage{
ukhls_clean_health_conditions(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Produce clean versions of indicator variables for health conditions.
}
