% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_smoke.R
\name{ukhls_clean_smoke}
\alias{ukhls_clean_smoke}
\title{Clean Smoking Variables}
\usage{
ukhls_clean_smoke(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all variables related to smoking behaviour - current smoking status, smoking history, and
level of cigarette consumption for current smokers.
}
