% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_work.R
\name{ukhls_clean_work}
\alias{ukhls_clean_work}
\title{Clean Work}
\usage{
ukhls_clean_work(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Generate variables (other than pay) relating to work, including hours of work for both
employed and self-employed, and industry / occupation.
}
