% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_combine_hhold_smokers.R
\name{ukhls_combine_hhold_smokers}
\alias{ukhls_combine_hhold_smokers}
\title{Read, Clean, Combine Understanding Society data}
\source{
University of Essex, Institute for Social and Economic Research. (2022). Understanding Society: Waves 1-12, 2009-2021 and Harmonised BHPS: Waves 1-18, 1991-2009. [data collection]. 17th Edition. UK Data Service. SN: 6614, \href{https://doi.org/10.5255/UKDA-SN-6614-18}{DOI: 10.5255/UKDA-SN-6614-18}
}
\usage{
ukhls_combine_hhold_smokers(data = NULL)
}
\arguments{
\item{data}{set of combined data}
}
\value{
Returns new set of combined data with smoking household counts
}
\description{
A wrapper function for applying all of the reading and cleaning functions, selecting the
desired variables/observations for the analysis, and specifying complete case
restrictions.
}
\details{
UKDS Study Number: \href{https://beta.ukdataservice.ac.uk/datacatalogue/studies/study?id=6614}{Understanding Society: Waves 1-12, 2009-2021 and Harmonised BHPS: Waves 1-18, 1991-2009}
}
