% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_post_process_benefits.R
\name{ukhls_post_clean_benefit}
\alias{ukhls_post_clean_benefit}
\title{Post process of Benefits variables}
\usage{
ukhls_post_clean_benefit(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced as part of the ukhls_clean_global function}
}
\description{
Produce benefits variables specifically for Unemployment (unemployed), In-work (employed) and Disability (inactive)
}
