% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_read_wave6.R
\name{ukhls_read_wave6}
\alias{ukhls_read_wave6}
\title{Read Understanding Society Wave 6}
\usage{
ukhls_read_wave6(
  root = c("C:/"),
  file = "Users/User/Documents/Datasets/UKHLS/tab/ukhls_w6/f_indresp.tab"
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}
}
\value{
Returns a data table. Note that:
\itemize{
\item Missing data ("NA", "", "-1", "-2", "-6", "-7", "-9", "-90", "-90.0", "N/A") is replace with NA,
except -8 ("don't know") as this is data.
\item All variable names are converted to lower case.
\item Each data point is assigned a weight of 1 as there is no weight variable supplied.
\item A single sampling cluster is assigned.
\item The probabilistic sampling unit have the year appended to them.
}
}
\description{
Reads and does basic cleaning on the UKHLS first wave.
}
\details{
A sample of the population living in private households. All persons living in the house, including those
under 2 years were eligible for inclusion. At addresses where there were more than two children under 16,
two children were selected at random. Information was obtained directly from persons aged 13 and
over. Information about children aged 0-12 was obtained from a parent, with the child present.

MISSING VALUES

\itemize{
\item -1 Don't know.
\item -2 Refused: Used only for variables on the nurse schedules, this code indicates that a
respondent refused a particular measurement or test or the measurement was attempted but not
obtained or not attempted.
\item -8 Not applicable: Used to signify that a particular variable did not apply to a given respondent
usually because of internal routing. For example, men in women only questions.
\item -9 Missing
}
}
