% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhlsclean_2020.R
\name{ukhlsclean_2020}
\alias{ukhlsclean_2020}
\title{Read Understanding Society 2020}
\source{
University of Essex, Institute for Social and Economic Research. (2022).
Understanding Society: Calendar Year Dataset, 2020. [data collection]. UK Data Service. SN: 8988, \href{https://doi.org/10.5255/UKDA-SN-8988-1}{DOI: 10.5255/UKDA-SN-8988-1}
}
\usage{
ukhlsclean_2020(
  root = c("X:/"),
  file = "HAR_PR/PR/USoc/Data/Calendar Year Datasets/SN8988_2022_11_29/tab",
  full = TRUE,
  ages = 16:89,
  country = "UK",
  keep_vars = NULL,
  complete_vars = c("d_age", "d_sex", "d_country")
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}

\item{full}{Logical - TRUE if restricting the sample to full interviews only (excluding proxies)}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{country}{Character - country to produce data for. One of c("UK","england","wales","scotland","northern_ireland"). Defaults to all UK.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year, age and sex).}
}
\value{
Returns a data table. Note that:
\itemize{
\item Missing data ("NA", "", "-1", "-2", "-6", "-7", "-9", "-90", "-90.0", "N/A") is replace with NA,
except -8 ("don't know") as this is data.
\item All variable names are converted to lower case.
\item Each data point is assigned a weight of 1 as there is no weight variable supplied.
\item A single sampling cluster is assigned.
\item The probabilistic sampling unit have the year appended to them.
}
}
\description{
Reads and cleans the Understanding Society calendar year data for 2020. These data are derived from the main
Understanding Society survey but designed to be a representative cross-section for the year 2020. For this data,
the function applies the full reading and cleaning proces.
}
\details{
UKDS Study Number: \href{https://beta.ukdataservice.ac.uk/datacatalogue/studies/study?id=8988}{SN8988 Understanding Society: Calendar Year Dataset, 2020}
}
