% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_econ_status.R
\name{clean_econ_status}
\alias{clean_econ_status}
\title{Clean Economic Status}
\usage{
clean_econ_status(data = NULL)
}
\description{
Generate three variables indicating economic status at different levels of detail; working vs not
working; employed, unemployed or inactive; and employed, self-employed, unemployed, sick, retired,
education, or other.
}
