% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_health.R
\name{clean_health}
\alias{clean_health}
\title{Clean Health and Caring Variables}
\usage{
clean_health(data = NULL)
}
\description{
Produce clean versions of variables measuring health outcomes. Also apply the
algorithm developed by Gray A, Rivero-Arias O, Clarke P (2006) to estimate
EQ-5D utility values from SF-12 responses
}
