% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_smoke.R
\name{clean_smoke}
\alias{clean_smoke}
\title{Clean Smoking Variables}
\usage{
clean_smoke(data = NULL)
}
\description{
Clean smoking variables and backwards impute smoking behaviour to waves prior to the smoking
questions being asked in the UKHLS.
}
