% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_smoke.R
\name{clean_smoke}
\alias{clean_smoke}
\title{Clean Smoking Variables}
\usage{
clean_smoke(data = NULL)
}
\description{
Reads and does basic cleaning on the UKHLS first wave.
}
\details{
A sample of the population living in private households. All persons living in the house, including those
under 2 years were eligible for inclusion. At addresses where there were more than two children under 16,
two children were selected at random. Information was obtained directly from persons aged 13 and
over. Information about children aged 0-12 was obtained from a parent, with the child present.
}
