% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select variables and apply filters}
\usage{
select_data(
  data,
  ages = NULL,
  country = NULL,
  complete_vars = NULL,
  calendar_year = FALSE
)
}
\arguments{
\item{data}{Data table - the cleaned Understanding Society dataset.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to NULL - all ages).}

\item{country}{Character - country to produce data for. One of c("england","wales","scotland","northern_ireland"). Defaults to NULL which includes all UK.}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to NULL - keep all observations).}

\item{calendar_year}{Logical - TRUE when the code is processing calendar year data (defaults to FALSE).}
}
\value{
Returns a reduced version of data
}
\description{
Selects the variables required for analysis and selects only the rows without missing data
for specified variables.
}
\examples{

\dontrun{


}

}
