% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhls_clean_hhold.R
\name{ukhls_clean_hhold}
\alias{ukhls_clean_hhold}
\title{Clean Household Variables}
\usage{
ukhls_clean_hhold(data = NULL)
}
\arguments{
\item{data}{Data table. Understanding Society data produced using the read functions.}
}
\description{
Clean all household level variables; household type (composition of adults and children),
household size, ownership status, and age of the youngest child in the household.
}
