% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhlsclean.R
\name{ukhlsclean}
\alias{ukhlsclean}
\title{Read, Clean, Combine Understanding Society data}
\source{
University of Essex, Institute for Social and Economic Research. (2022). Understanding Society: Waves 1-12, 2009-2021 and Harmonised BHPS: Waves 1-18, 1991-2009. [data collection]. 17th Edition. UK Data Service. SN: 6614, \href{https://doi.org/10.5255/UKDA-SN-6614-18}{DOI: 10.5255/UKDA-SN-6614-18}
}
\usage{
ukhlsclean(
  root = "X:/",
  file = "HAR_PR/PR/USoc/Data/SN6614_2022_11_29/tab/ukhls",
  full = TRUE,
  waves = 1:12,
  ages = 16:89,
  keep_vars = NULL,
  complete_vars = c("d_age", "d_sex")
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}

\item{full}{Logical - TRUE if restricting the sample to full interviews only (excluding proxies)}

\item{waves}{Integer vector - the waves of the UKHLS to retain (defaults to all - 1 to 11).}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year, age and sex).}
}
\value{
Returns a new set of variables
}
\description{
A wrapper function for applying all of the reading and cleaning functions, selecting the
desired variables/observations for the analysis, and specifying complete case
restrictions.
}
\details{
UKDS Study Number: \href{https://beta.ukdataservice.ac.uk/datacatalogue/studies/study?id=6614}{Understanding Society: Waves 1-12, 2009-2021 and Harmonised BHPS: Waves 1-18, 1991-2009}
}
